short instead of byte

+++ Plugin Reverse Socks Common/src/se/jrat/plugins/socks/Global.java
@@ -2,7 +2,7 @@ package se.jrat.plugins.socks;
 
 public class Global {
 	
-	public static final byte HEADER_START = 118;
-	public static final byte HEADER_STOP = 119;
+	public static final short HEADER_START = 118;
+	public static final short HEADER_STOP = 119;
 
 }
